pyproject.toml
[tool.poetry]
name = "ferry"
version = "0.1.0"
description = ""
authors = ["Dan Trickey <git@trickey.io>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
django-stubs-ext = "^4.2.7"
django-ninja = "^1.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-django = "^4.8.0"
djhtml = "^3.0.6"
mypy = "^1.7.0"
django-coverage-plugin = "^3.1.0"
django-stubs = {version = "^4.2.7", extras = ["compatible-mypy"]}
factory-boy = "^3.3.0"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "A",
    "ANN",
    "B",
    "BLE",
    "C4",
    "COM",
    "DJ",
    "DTZ",
    "E",
    "F",
    "FA",
    "FBT",
    "I",
    "N",
    "S",
    "T10",
    "UP",
    "W",
]
ignore = [
    "ANN101",  # Missing type annotation for `self` in method
    "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed
    "B009",    # Do not call `getattr` with a constant attribute value.
    "COM812",  # Conflicts with formatter
    "N999",    # Invalid module name
    "S101",    # S101 Use of `assert` detected
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

warn_unused_ignores = true

strict_optional = true
no_implicit_optional = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true

check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["factory.*",]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["*.tests.*"]
disable_error_code = ["arg-type"]

[tool.django-stubs]
django_settings_module = "ferry.core.settings"

[tool.coverage.run]
plugins = [
    'django_coverage_plugin',
]

omit = [
    "ferry/wsgi.py",
    "ferry/asgi.py",
    "*/migrations/*",
    "*/tests/*",
]

[tool.coverage.django_coverage_plugin]
template_extensions = 'html, txt, tex, email'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"